+++ src/components/Sidebar/NavSection.tsx
@@ -1,0 +1,18 @@
+import { Box, Stack, Text } from "@chakra-ui/react";
+import { ReactNode } from "react";
+
+interface NavSectionPros {
+    title: string;
+    children: ReactNode;
+}
+
+export function NavSection({ title, children }: NavSectionPros) {
+    return (
+        <Box>
+            <Text fontWeight="bold" color="gray.400" fontSize="small">{title}</Text>
+            <Stack spacing="4" mt="8" align="strech">
+                {children}
+            </Stack>
+        </Box>
+    );
+}
